+++ core/src/main/java/dev/olog/core/interactor/playlist/RemoveDuplicatesUseCase.kt
@@ -15,8 +15,8 @@ class RemoveDuplicatesUseCase @Inject constructor(
 
     suspend operator fun invoke(mediaId: Category) {
         return when (mediaId.category){
-            PODCASTS_PLAYLIST -> podcastPlaylistGateway.removeDuplicated(mediaId.categoryId)
-            PLAYLISTS -> playlistGateway.removeDuplicated(mediaId.categoryId)
+            PODCASTS_PLAYLIST -> podcastPlaylistGateway.removeDuplicated(mediaId.categoryId.toLong())
+            PLAYLISTS -> playlistGateway.removeDuplicated(mediaId.categoryId.toLong())
             else -> throw IllegalArgumentException("invalid media id $mediaId")
         }
     }
